Separate helix-term as a library
helix-term stuff will now be documented in rustdoc.
